pyproject.toml
[project]
name = "screamsheet-dispatch"
version = "0.1.0"
description = "Automated delivery system for screamsheet PDFs"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "gspread>=6.0",
    "google-auth>=2.0",
    "sqlalchemy>=2.0",
]

[project.scripts]
screamsheet-dispatch = "screamsheet_dispatch.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/screamsheet_dispatch"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
]
